Add half working vectorized replace solution
The code is cleaner than the last replace-too-big approach , but it's
just as slow (or slower!). Each loop adds approx 4 seconds to runtime,
so it'll have to run 9 times for 99.95% of the values to be replaced
((1 - (.25 **9)) ** 128) in the warp. That's slower than my original
solution with modulus.

This code doesn't work, but it demonstrates performance. There's
likely a masking error.